Fixed bug issue with glow sprites not displaying correctly

+++ src/smallExplosion.hpp
@@ -26,7 +26,7 @@ class SmallExplosion : public Effect {
 private:
 	mutable SpriteSheet<174, 224, 36, 36> spriteSheet;
 	sf::Sprite glow;
-	int32_t glowFadeTimer;
+	int64_t glowFadeTimer;
 	
 public:
 	SmallExplosion(const sf::Texture &, const sf::Texture &, float, float);
